fix: separate club registration module
