+++ packages/api/src/feature/registration/registration.module.ts
@@ -2,21 +2,15 @@ import { Module } from "@nestjs/common";
 
 import { DrizzleModule } from "src/drizzle/drizzle.module";
 
-import { ClubModule } from "../club/club.module";
-// import ClubPublicService from "../club/service/club.public.service";
-
+import { ClubRegistrationModule } from "./club-registration/club-registration.module";
 import { RegistrationController } from "./controller/registration.controller";
 import { MemberRegistrationModule } from "./member-registration/member-registration.module";
 import { RegistrationRepository } from "./repository/registration.repository";
 import { RegistrationService } from "./service/registration.service";
 
 @Module({
-  imports: [DrizzleModule, ClubModule, MemberRegistrationModule],
+  imports: [DrizzleModule, MemberRegistrationModule, ClubRegistrationModule],
   controllers: [RegistrationController],
-  providers: [
-    RegistrationRepository,
-    RegistrationService,
-    MemberRegistrationModule,
-  ],
+  providers: [RegistrationRepository, RegistrationService],
 })
 export class RegistrationModule {}
